API/requirements.txt
aiocsv==1.2.3
aiofiles==23.1.0
aioredis==2.0.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.6.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.5.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
